backend/requirements.txt
wheel
fastapi
fastapi-utils
python-arango
gunicorn
uvicorn
ujson
pydantic
python-multipart
aiofiles